+++ src/context/deck-creator/deck-creator.action-creators.ts
@@ -1,9 +1,11 @@
 import {
   DeckCreatorAction,
   SET_DESCRIPTION,
-  SET_IMAGE_STEP_DONE,
   SET_IMAGE_URL,
   SET_NAME,
+  SET_READY_TO_SUBMIT,
+  SET_STEP,
+  SET_TAGS,
   START_UPLOADING_IMAGE,
   STOP_UPLOADING_IMAGE,
 } from './deck-creator.reducer';
@@ -21,13 +23,6 @@ export const stopUploadingDeckImage: () => DeckCreatorAction = () => ({
   type: STOP_UPLOADING_IMAGE,
 });
 
-export const setDeckImageStepDone: (imageStepDone: boolean) => DeckCreatorAction = (
-  imageStepDone,
-) => ({
-  type: SET_IMAGE_STEP_DONE,
-  imageStepDone,
-});
-
 export const setDeckName: (name: string) => DeckCreatorAction = (name) => ({
   type: SET_NAME,
   name,
@@ -36,4 +31,21 @@ export const setDeckName: (name: string) => DeckCreatorAction = (name) => ({
 export const setDeckDescription: (description: string) => DeckCreatorAction = (description) => ({
   type: SET_DESCRIPTION,
   description,
+});
+
+export const setDeckCreatorStep: (step: number) => DeckCreatorAction = (step) => ({
+  type: SET_STEP,
+  step,
+});
+
+export const setDeckTags: (tags: string[]) => DeckCreatorAction = (tags) => ({
+  type: SET_TAGS,
+  tags,
+});
+
+export const setDeckReadyToSubmit: (isReadyToSubmit: boolean) => DeckCreatorAction = (
+  isReadyToSubmit,
+) => ({
+  type: SET_READY_TO_SUBMIT,
+  isReadyToSubmit,
 });