+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-# 0.1.0 (2020-10-10)
+# 0.1.0 (2020-10-16)
 
 
 ### Features
 
 - Added locale feature ([a4d998c](https://github.com/Miczeq22/brainflash-app/commit/a4d998c20b3fad2348b53b636652adc5e95f3f4d))
+- Deck creator page ([061d52c](https://github.com/Miczeq22/brainflash-app/commit/061d52c03a3fc4edb7f36cc02e89de5d82046055))
 - Deck details page ([c826fa2](https://github.com/Miczeq22/brainflash-app/commit/c826fa28b21aa21c42397604764f0aa090d00cdf))
+- Deck image uploader and deck creator context ([2be5a58](https://github.com/Miczeq22/brainflash-app/commit/2be5a58326442c6be1b145b5094f2b1dbec62da7))
 - Deck List Component ([e88d6cd](https://github.com/Miczeq22/brainflash-app/commit/e88d6cdcdbffb18d95777fba4d3c2f6f9a27909b))
 - Deck list item component ([86759ae](https://github.com/Miczeq22/brainflash-app/commit/86759aebacc500f250cae836ec33c84634596d89))
 - Deck List Provider ([5ee9ed5](https://github.com/Miczeq22/brainflash-app/commit/5ee9ed5b0b26db8e293466427fb2597f863df185))
@@ -17,6 +19,7 @@
 - Menu component ([5dfffc1](https://github.com/Miczeq22/brainflash-app/commit/5dfffc1648899f3a0d9d6694919dd6df6b96df58))
 - No deck selected component ([37ffb73](https://github.com/Miczeq22/brainflash-app/commit/37ffb73b45d1be09a761c65f9d8adc71fb2e5673))
 - Ratings component ([f1b89ef](https://github.com/Miczeq22/brainflash-app/commit/f1b89ef38a4696bc80384e230d51657d19b78537))
+- Refresh token interceptor ([33eef08](https://github.com/Miczeq22/brainflash-app/commit/33eef0872b2c3b0d9b0c2e5a66456c91579c6af7))
 - Register page ([6bb27ae](https://github.com/Miczeq22/brainflash-app/commit/6bb27ae069b1e15db8b073c92c4b7809df61f2dd))
 - Topbar component ([7e5c376](https://github.com/Miczeq22/brainflash-app/commit/7e5c3767550fe1478d82d0699041192a790e12fe))
 
@@ -27,7 +30,7 @@
 - Auth fetch for apollo client ([a68dcd6](https://github.com/Miczeq22/brainflash-app/commit/a68dcd6c7902a4db34c6f406b6ac7de9411c3b49))
 - Connected local storage for auth provider ([f483437](https://github.com/Miczeq22/brainflash-app/commit/f4834372c131d2706ea2d1d99f6060b908b2d568))
 - Empty list improvement ([19f629c](https://github.com/Miczeq22/brainflash-app/commit/19f629c3ed7140f1b598eb8a4eaefc956ac10e4e))
-- Get user data from JWT ([d29132a](https://github.com/Miczeq22/brainflash-app/commit/d29132a8862fdd3364455f285ab5e17a73c035c0))
+- Get user data from JWT ([07ce790](https://github.com/Miczeq22/brainflash-app/commit/07ce79002d82d7b1b13f097805cba7b18ad0104f))
 - Loader for decks ([934bda3](https://github.com/Miczeq22/brainflash-app/commit/934bda360a452d380c98ea551a530635b5665b37))
 - Refactored deck list and deck list item ([992bd25](https://github.com/Miczeq22/brainflash-app/commit/992bd25b6988955b343d829afc8fc902edd80504))